frontend/requirements.txt
gradio
python-dotenv
requests